+++ resources/views/tenantarea/pages/index.blade.php
@@ -8,11 +8,11 @@
 
 @section('body-attributes')data-spy="scroll" data-offset="0" data-target="#navigation"@endsection
 
-@section('meta-title', $currentTenant->name)
-@section('meta-description', strip_tags($currentTenant->description))
+@section('meta-title', app('request.tenant')->name)
+@section('meta-description', strip_tags(app('request.tenant')->description))
 
-@if($currentTenant->getFirstMediaUrl('cover_photo'))
-    @section('meta-image', $currentTenant->getFirstMediaUrl('cover_photo'))
+@if(app('request.tenant')->getFirstMediaUrl('cover_photo'))
+    @section('meta-image', app('request.tenant')->getFirstMediaUrl('cover_photo'))
 @endif